BatterStatsAPI

+++ migrations/20201220165707-create-standings.js
@@ -39,7 +39,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       pct: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(4,3)
       },
       gb: {
         type: Sequelize.CHAR
